Refactor pg_upgrade and gpstop status checking
Previously, the code in the upgradestatus package to check gpstop status was
very similar to the code for checking pg_upgrade status, one of the two having
largely been copy-pasted from the other.

This commit unifies and genericizes the status check logic in shutdown_clusters
and pg_upgrade_status_checker and merges the two files into a new file named
utility_status_checker.

There are minor logic changes as well (e.g.  refactoring isUpgradeComplete for
simplicity), but overall the functionality is unchanged, and there are no code
changes outside of the upgradestatus package.

Authored-by: Jamie McAtamney <jmcatamney@pivotal.io>
